// v86/libv86.d.ts
	type V86Image =
		| { url: string }
		| { buffer: ArrayBuffer }
		| { url: string; async: false; size?: number }
		| { url: string; async: true; size: number };

	/**
	 * A 9p filesystem is supported by the emulator, using a virtio transport. Using it, files can be exchanged with the guest OS
	 * If `basefs` and `baseurl` are omitted, an empty 9p filesystem is created.
	 */
	interface Filesystem9pOptions {
		/**
		 * json file created using [fs2json](https://github.com/copy/v86/blob/master/tools/fs2json.py).
		 */
		baseurl?: string;

		/**
		 * The base url is the prefix of a url from which the files are available.
		 * For instance, if the 9p filesystem has a file `/bin/sh`, that file must be accessible from http://localhost/9p/base/bin/sh
		 */
		basefs?: string;
	}

	enum LogLevel {
		LOG_ALL = -1,
		LOG_NONE = 0,
		LOG_OTHER = 0x000001,
		LOG_CPU = 0x000002,
		LOG_FPU = 0x000004,
		LOG_MEM = 0x000008,
		LOG_DMA = 0x000010,
		LOG_IO = 0x000020,
		LOG_PS2 = 0x000040,
		LOG_PIC = 0x000080,
		LOG_VGA = 0x000100,
		LOG_PIT = 0x000200,
		LOG_MOUSE = 0x000400,
		LOG_PCI = 0x000800,
		LOG_BIOS = 0x001000,
		LOG_FLOPPY = 0x002000,
		LOG_SERIAL = 0x004000,
		LOG_DISK = 0x008000,
		LOG_RTC = 0x010000,
		LOG_HPET = 0x020000,
		LOG_ACPI = 0x040000,
		LOG_APIC = 0x080000,
		LOG_NET = 0x100000,
		LOG_VIRTIO = 0x200000,
		LOG_9P = 0x400000,
		LOG_SB16 = 0x800000,
	}

	enum BootOrder {
		CD_FLOPPY_HARDDISK = 0x213,
		CD_HARDDISK_FLOPPY = 0x123,
		FLOPPY_CD_HARDDISK = 0x231,
		FLOPPY_HARDDISK_CD = 0x321,
		HARDDISK_CD_FLOPPY = 0x132,
	}
// ...
	interface V86Options {
		/**
		 * Path to v86 wasm artifact
		 * @default "build/v86.wasm" or "build/v86-debug.wasm" when debug mode enabled
		 */
		wasm_path?: string;

		/**
		 * The memory size in bytes, should be a power of 2.
		 * @example 16 * 1024 * 1024
		 * @default 64 * 1024 * 1024
		 */
		memory_size?: number;

		/**
		 * VGA memory size in bytes.
		 * @example 8 * 1024 * 1024
		 * @default 8 * 1024 * 1024
		 */
		vga_memory_size?: number;

		/**
		 * If emulation should be started when emulator is ready.
		 * @default false
		 */
		autostart?: boolean;

		/**
		 * If keyboard should be disabled.
		 * @default false
		 */
		disable_keyboard?: boolean;

		/**
		 * If mouse should be disabled.
		 * @default false
		 */
		disable_mouse?: boolean;

		/**
		 * If speaker should be disabled.
		 * @default false
		 */
		disable_speaker?: boolean;

		/**
		 * The url of a server running websockproxy. See [networking.md](networking.md). Setting this will enable an emulated network card.
		 * @default undefined
		 */
		network_relay_url?: string;

		/**
		 * Either a url pointing to a bios or an ArrayBuffer, see below.
		 * @default undefined
		 */
		bios?: V86Image;

		/**
		 * VGA bios, see below.
		 * @default undefined
		 */
		vga_bios?: V86Image;

		/**
		 * First hard disk, see below.
		 * @default undefined
		 */
		hda?: V86Image;

		/**
		 * First floppy disk, see below.
		 * @default undefined
		 */
		fda?: V86Image;

		/**
		 * cdrom
		 * @default undefined
		 */
		cdrom?: V86Image;

		/**
		 * A Linux kernel image to boot (only bzimage format)
		 * @default undefined
		 */
		bzimage?: V86Image;

		/**
		 * A Linux ramdisk image
		 * @default undefined
		 */
		initrd?: V86Image;

		/**
		 * Automatically fetch bzimage and initrd from the specified `filesystem`.
		 */
		bzimage_initrd_from_filesystem?: boolean;

		/**
		 * An initial state to load
		 * @default undefined
		 */
		initial_state?: V86Image;

		/**
		 * A 9p filesystem
		 * @default undefined
		 */
		filesystem?: Filesystem9pOptions;

		/**
		 * A textarea that will receive and send data to the emulated serial terminal.
		 * Alternatively the serial terminal can also be accessed programatically, see [serial.html](../examples/serial.html).
		 * @default undefined
		 */
		serial_container?: HTMLTextAreaElement;

		/**
		 * Xtermjs serial terminal container. When set, serial_container option is ignored.
		 * @default undefined
		 */
		serial_container_xtermjs?: HTMLElement;

		/**
		 * An HTMLElement. This should have a certain structure, see [basic.html](../examples/basic.html).
		 * @default undefined
		 */
		screen_container?: HTMLElement | null;

		/**
		 * ACPI
		 * @default false
		 */
		acpi?: boolean;

		/**
		 * log level
		 * @default LogLevel.LOG_NONE
		 */
		log_level?: LogLevel;

		/**
		 * boot order
		 * @default BootOrder.CD_FLOPPY_HARDDISK
		 */
		boot_order?: BootOrder;

		/**
		 * fast boot
		 * @default false
		 */
		fastboot?: boolean;

		/**
		 * enables UART1 (Serial port)
		 * @see http://wiki.osdev.org/UART
		 * @see https://github.com/s-macke/jor1k/blob/master/js/worker/dev/uart.js
		 * @see https://www.freebsd.org/doc/en/articles/serial-uart/
		 * @default undefined
		 */
		uart1?: boolean;

		/**
		 * enables UART2 (Serial port)
		 * @see http://wiki.osdev.org/UART
		 * @see https://github.com/s-macke/jor1k/blob/master/js/worker/dev/uart.js
		 * @see https://www.freebsd.org/doc/en/articles/serial-uart/
		 * @default undefined
		 */
		uart2?: boolean;

		/**
		 * enables UART3 (Serial port)
		 * @see http://wiki.osdev.org/UART
		 * @see https://github.com/s-macke/jor1k/blob/master/js/worker/dev/uart.js
		 * @see https://www.freebsd.org/doc/en/articles/serial-uart/
		 * @default undefined
		 */
		uart3?: boolean;

		/**
		 * boot cmdline
		 */
		cmdline?: string;

		/**
		 * Ne2k: should MAC be preserved from the state image
		 * @default undefined
		 */
		preserve_mac_from_state_image?: boolean;

		/**
		 * custom network adapter
		 * @default undefined
		 */
		network_adapter?: NetworkAdapter;

		/**
		 * enables screen dummy
		 * @default undefined
		 */
		screen_dummy?: boolean;

		mac_address_translation?: boolean;
	}